renames import paths for SmallFormButton and adds Save & Quit button to step2.js
