+++ src/pages/UserSettings/AccountInfo/Step4.js
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 import { useForm, Controller } from "react-hook-form";
 import { useParams, useNavigate } from "@reach/router";
 // Component imports
-import NextButton from "../../../theme/FormButton";
+import NextButton from "../../../theme/SmallFormButton";
 import ProgressBar from "../../../theme/ProgressBar";
 // Material-UI imports
 import {
