+++ src/pages/UserSettings/AccountInfo/Step6.js
@@ -3,7 +3,7 @@ import React from "react";
 import { useForm, Controller } from "react-hook-form";
 import { useParams, useNavigate } from "@reach/router";
 // Component imports
-import FinishButton from "../../../theme/FormButton";
+import FinishButton from "../../../theme/SmallFormButton";
 import ProgressBar from "../../../theme/ProgressBar";
 // Material-UI imports
 import {
